chore: Initialize Advent of Code Repository
Created a new GitHub repository for my Advent of Code solutions. The README provides an overview of the challenge, my progress, and a contribution notice. Happy coding and may the stars be with you!

The planned Structure will result in a separate directory per AoC Event (Year), with a single Markdown File per Day Completed.